Added anonymizing feature for sensitive info.
Anonymizing can be requested either from the command line by using
an option '--anonymize, -n' or set to default in the config file.

When activated it applies equally to both the console output and
the JSON output file.

The underlying implementation uses Blake2b hash (96-bit), which
hashes: <manufacturer name>|<model name>|<actual S/N>
and outputs the string in the form "anon@<hash>" instead of the
original info.