Changed: Total vertical FOV calculation formula.
The total vertical FOV is now calculated as an average of the left and right vertical FOVs, instead of using combined minimum.

The fix has been added to `hmdv` which will recalculate the FOV and save the correct values.

Some parts of `jtools` were spun out into a separate module.